# MQE: Refactor functions over instant vectors to support scalar args and add clamp function