--- a/pkg/streamingpromql/operators/function_over_instant_vector.go
+++ b/pkg/streamingpromql/operators/function_over_instant_vector.go
@@ -70,7 +70,7 @@ func (m *FunctionOverInstantVector) processScalarArgs(ctx context.Context) error
 }
 
 func (m *FunctionOverInstantVector) SeriesMetadata(ctx context.Context) ([]types.SeriesMetadata, error) {
-	// Pre-process any Scalar agruments
+	// Pre-process any Scalar arguments
 	err := m.processScalarArgs(ctx)
 	if err != nil {
 		return nil, err