--- a/pkg/streamingpromql/operators/function_over_instant_vector_test.go
+++ b/pkg/streamingpromql/operators/function_over_instant_vector_test.go
@@ -38,7 +38,7 @@ func TestFunctionOverInstantVector(t *testing.T) {
 	}
 
 	seriesDataFuncCalledTimes := 0
-	mustBeCalledSeriesData := func(types.InstantVectorSeriesData, *limiting.MemoryConsumptionTracker) (types.InstantVectorSeriesData, error) {
+	mustBeCalledSeriesData := func(types.InstantVectorSeriesData, []types.ScalarData, *limiting.MemoryConsumptionTracker) (types.InstantVectorSeriesData, error) {
 		seriesDataFuncCalledTimes++
 		return types.InstantVectorSeriesData{}, nil
 	}