--- a/pkg/streamingpromql/testdata/ours-only/native_histograms.test
+++ b/pkg/streamingpromql/testdata/ours-only/native_histograms.test
@@ -1,0 +1,9 @@
+# Test metric with mixed floats and histograms
+load 1m
+	mixed_metric {{schema:0 sum:5 count:4 buckets:[1 2 1]}} 1 2 3 {{schema:0 sum:5 count:4 buckets:[1 2 1]}} {{schema:0 sum:8 count:6 buckets:[1 4 1]}}
+
+# clamp ignores any histograms
+# Prometheus currently returns min instead of no value as per the documentation
+# https://github.com/prometheus/prometheus/pull/15169
+eval range from 0 to 5m step 1m clamp(mixed_metric, 2, 5)
+	{} _ 2 2 3