--- a/pkg/streamingpromql/testdata/ours/functions.test
+++ b/pkg/streamingpromql/testdata/ours/functions.test
@@ -404,3 +404,13 @@ eval_fail instant at 5m tan({__name__=~"float_metric_.*"})
 
 eval_fail instant at 5m tanh({__name__=~"float_metric_.*"})
   expected_fail_message vector cannot contain metrics with the same labelset
+
+clear
+
+load 5m
+  mins   0   10 0  30 0   -10  0  5   NaN 0   NaN NaN
+  maxes  10  20 5  0  -10 -5   40 50  10  NaN NaN NaN
+  series -10 0  10 20 30  40   50 NaN 60  70  80  NaN
+
+eval range from 0 to 60m step 5m clamp(series, scalar(mins), scalar(maxes))
+  {}     0   10 5  _  _   -5   40 NaN NaN NaN NaN NaN NaN